feat(create-project): Added logic to create project into the form.
Created hook usePostProject which returns a function to make the post request, isError, errorMessage and isLoading. This hook is already added into the form. Error Message handeling and Loading message is not enabled.

BREAKING CHANGE: path on service postProject was changed to "" instead of `Projects/`. When needed please change it back.